requirements.txt
absl-py==0.12.0
appdirs==1.4.4
bandit==1.7.0
beautifulsoup4==4.9.3
black==20.8b1
click==7.1.2
environs==9.3.1
gitdb==4.0.5
GitPython==3.1.14
marshmallow==3.10.0
mypy-extensions==0.4.3
pathspec==0.8.1
pbr==5.5.1
python-dotenv==0.15.0
PyYAML==5.4.1
regex==2021.3.17
six==1.15.0
smmap==3.0.5
soupsieve==2.2.1
stevedore==3.3.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
